move vendored sqlite3 files to its own directory
