+++ src/database/sqlite/sqlite3/README.md
@@ -1,0 +1,1 @@
+Vendored copy of SQLITE3
