+++ src/database/sqlite/sqlite3/dbdata.c
