+++ src/database/sqlite/sqlite_functions.h
@@ -4,7 +4,7 @@
 #define NETDATA_SQLITE_FUNCTIONS_H
 
 #include "daemon/common.h"
-#include "sqlite3.h"
+#include "database/sqlite/sqlite3/sqlite3.h"
 
 void analytics_set_data_str(char **name, const char *value);
